:hammer: Update dev knex migrate make command to work with docker compose watch sync setup.
Now that we aren't using volumes, we don't get a two way sync of the migrations directory.
On the plus side, a docker compose cp handles files permissions so we no longer need to run as sudo and take over the file.

NOTE: this also means that we need to use "docker compose exec" rather than "docker compose run" for all migration commands, as the migration file won't exist inside the temporary container created by "run".